Speed-optimized version of split_pkg_name
Instead of a command substitution with 'echo' and 'sed', which comes at the expense of launching a subshell plus an external command, this version uses parameter substitution, which is functionality built in to the shell. Even if a little more verbose, it should be about two orders of magnitude faster.